Remove obsolete test project and related tests for BaseKubernetesGenerator
