--- Devantler.KubernetesGenerator.Core.Tests/omit-defaults.verified.yaml
@@ -1,2 +1,0 @@
-﻿---
-name: my-name
