+++ Devantler.KubernetesGenerator.K3d/K3dConfigGenerator.cs
@@ -6,6 +6,6 @@ namespace Devantler.KubernetesGenerator.K3d;
 /// <summary>
 /// Generator for generating K3d config files.
 /// </summary>
-public class K3dConfigGenerator : BaseKubernetesGenerator<K3dConfig>
+public class K3dConfigGenerator() : BaseKubernetesGenerator<K3dConfig>(omitDefaults: true)
 {
 }
