+++ Devantler.KubernetesGenerator.Kind.Tests/KindConfigGeneratorTests/kind-config.full.verified.yaml
@@ -2,32 +2,23 @@
 kind: Cluster
 apiVersion: kind.x-k8s.io/v1alpha4
 name: kind-default
-featureGates:
-  CSIMigration: false
+featureGates: {}
 runtimeConfig:
   api/alpha: false
 networking:
-  ipFamily: ipv4
   apiServerAddress: 127.0.0.1
   apiServerPort: 6443
   podSubnet: 10.244.0.0/16
   serviceSubnet: 10.96.0.0/16
-  disableDefaultCNI: false
-  kubeProxyMode: iptables
 nodes:
-- role: ControlPlane
-  image: kindest/node:v1.32.0
+- image: kindest/node:v1.32.0
   extraMounts:
   - hostPath: /var/lib/kubelet
     containerPath: /var/lib/kubelet
-    readOnly: false
-    selinuxRelabel: false
-    propagation: None
   extraPortMappings:
   - hostPort: 6443
     containerPort: 6443
     listenAddress: 0.0.0.0
-    protocol: TCP
   kindNodeLabels:
     node-role.kubernetes.io/master: ''
   kubeadmConfigPatches: ''
@@ -36,14 +27,10 @@ nodes:
   extraMounts:
   - hostPath: /var/lib/kubelet
     containerPath: /var/lib/kubelet
-    readOnly: false
-    selinuxRelabel: false
-    propagation: None
   extraPortMappings:
   - hostPort: 6443
     containerPort: 6443
     listenAddress: 0.0.0.0
-    protocol: TCP
   kindNodeLabels:
     node-role.kubernetes.io/worker: ''
   kubeadmConfigPatches: ''
