+++ Devantler.KubernetesGenerator.Kind/KindConfigGenerator.cs
@@ -6,6 +6,6 @@ namespace Devantler.KubernetesGenerator.Kind;
 /// <summary>
 /// Generator for generating K3d config files.
 /// </summary>
-public class KindConfigGenerator : BaseKubernetesGenerator<KindConfig>
+public class KindConfigGenerator() : BaseKubernetesGenerator<KindConfig>(omitDefaults: true)
 {
 }
